+++ R/DiffusionTF.R
@@ -47,8 +47,8 @@
 ##' }
 ##' @author Thomas Münch
 ##' @references Münch, T. and Laepple, T.: What climate signal is contained in
-##' decadal to centennial scale isotope variations from Antarctic ice cores?
-##' Clim. Past Discuss., https://doi.org/10.5194/cp-2018-112, in review, 2018.
+##' decadal- to centennial-scale isotope variations from Antarctic ice cores?
+##' Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
 ##' @export
 DiffusionTF <- function(nt, nc, ns, sigma, res = 1, coherent = FALSE, ...) {
 
@@ -79,9 +79,9 @@ DiffusionTF <- function(nt, nc, ns, sigma, res = 1, coherent = FALSE, ...) {
     for (i in 1 : ns) {
 
         if (coherent) {
-            X <- array(rnorm(nt), dim = c(nt, nc))
+            X <- array(stats::rnorm(nt), dim = c(nt, nc))
         } else {
-            X <- array(rnorm(nt * nc), dim = c(nt, nc))
+            X <- array(stats::rnorm(nt * nc), dim = c(nt, nc))
         }
         
         Xdiff <- array(dim = c(nt, nc))
@@ -100,10 +100,10 @@ DiffusionTF <- function(nt, nc, ns, sigma, res = 1, coherent = FALSE, ...) {
     # calculate spectra of the 'ns' average diffused and undiffused noise series
 
     signal.spec <- lapply(seq_len(ncol(stack.signal)), function(i) {
-        SpecMTM(ts(stack.signal[, i], deltat = res), ...)})
+        SpecMTM(stats::ts(stack.signal[, i], deltat = res), ...)})
 
     diff.spec <- lapply(seq_len(ncol(stack.diff)), function(i) {
-        SpecMTM(ts(stack.diff[, i], deltat = res), ...)})
+        SpecMTM(stats::ts(stack.diff[, i], deltat = res), ...)})
 
 
     # calculate the average over all 'ns' simulations