+++ R/TimeUncertaintyTF.R
@@ -17,7 +17,7 @@
 ##' is available on GitHub under
 ##' \url{https://github.com/EarthSystemDiagnostics/simproxyage}
 ##' and can be installed directly using
-##' \code{devtools::install_github("EarthSystemDiagnostics/simproxyage")}.
+##' \code{remotes::install_github("EarthSystemDiagnostics/simproxyage")}.
 ##'
 ##' The spectral estimates are calculated using Thomson’s multitaper method with
 ##' three windows with linear detrending before analysis.
@@ -98,16 +98,17 @@
 ##' Comboul, M., Emile-Geay, J., Evans, M. N., Mirnateghi, N., Cobb, K. M.
 ##' and Thompson, D. M.: A probabilistic model of chronological errors in
 ##' layer-counted climate proxies: applications to annually banded coral
-##' archives, Clim. Past, 10(2), 825-841, doi: 10.5194/cp-10-825-2014, 2014.
+##' archives, Clim. Past, 10(2), 825-841,
+##' https://doi.org/10.5194/cp-10-825-2014, 2014.
 ##'
 ##' Münch, T. and Laepple, T.: What climate signal is contained in
-##' decadal to centennial scale isotope variations from Antarctic ice cores?
-##' Clim. Past Discuss., https://doi.org/10.5194/cp-2018-112, in review, 2018.
+##' decadal- to centennial-scale isotope variations from Antarctic ice cores?
+##' Clim. Past, 14, 2053–2070, https://doi.org/10.5194/cp-14-2053-2018, 2018.
 ##' @export
 TimeUncertaintyTF <- function(t = 100 : 1, acp = c(t[1], NA),
                               nt = length(t), nc = 1, ns = 100,
                               model = "poisson", rate = 0.05, resize = 1,
-                              surrogate.fun = rnorm, fun.par = NULL,
+                              surrogate.fun = stats::rnorm, fun.par = NULL,
                               pad = TRUE, ...) {
 
     # check if package simproxyage is available
@@ -144,10 +145,10 @@ TimeUncertaintyTF <- function(t = 100 : 1, acp = c(t[1], NA),
     # calculate spectra of the 'ns' input and age-perturbed time series
 
     stacks.spec <- lapply(seq_len(ncol(stacks)), function(i) {
-        SpecMTM(ts(stacks[, i], deltat = 1), ...)})
+        SpecMTM(stats::ts(stacks[, i], deltat = 1), ...)})
 
     input.spec <- lapply(seq_len(ncol(run$input)), function(i) {
-        SpecMTM(ts(run$input[, i], deltat = 1), ...)})
+        SpecMTM(stats::ts(run$input[, i], deltat = 1), ...)})
 
 
     # calculate the average over all 'ns' simulations