+++ data-raw/clim-site-par.R
@@ -3,7 +3,7 @@
 ## Author: Thomas Münch (thomas.muench@awi.de), Alfred-Wegener-Institut, 2018
 ##
 
-library(devtools)
+library(usethis)
 
 dml <- read.table("data-raw/dml_diffusion-par.txt",
                   header = TRUE, skip = 10, sep = "\t")
@@ -12,5 +12,5 @@ wais <- read.table("data-raw/wais_diffusion-par.txt",
 
 clim.site.par <- rbind(dml, wais)
 
-devtools::use_data(clim.site.par, overwrite = TRUE)
+usethis::use_data(clim.site.par, overwrite = TRUE)