+++ data-raw/load.Trench.R
@@ -3,7 +3,7 @@
 ## Author: Thomas Münch (thomas.muench@awi.de), Alfred-Wegener-Institut, 2018
 ##
 
-library(devtools)
+library(usethis)
 
 # 'TrenchR' package to load the data
 library(TrenchR)
@@ -27,5 +27,5 @@ t15 <- as.data.frame(cbind(t15.1, t15.2))
 # Document time of processing
 attr(t15, "info") <- sprintf("Processed on: %s.", Sys.time())
 
-devtools::use_data(t15, overwrite = TRUE)
+usethis::use_data(t15, overwrite = TRUE)